# Refactor common.TestCase: speed up tests