--- a/tests/countries/test_poland.py
+++ b/tests/countries/test_poland.py
@@ -16,7 +16,8 @@ from tests.common import TestCase
 class TestPoland(TestCase):
     @classmethod
     def setUpClass(cls):
-        super().setUpClass(Poland)
+        years = range(1925, 2050)
+        super().setUpClass(Poland, years=years)
 
     def test_country_aliases(self):
         self.assertCountryAliases(Poland, PL, POL)
@@ -27,6 +28,7 @@ class TestPoland(TestCase):
     def test_2017(self):
         # http://www.officeholidays.com/countries/poland/2017.php
         self.assertHolidayDates(
+            Poland(years=2017),
             "2017-01-01",
             "2017-01-06",
             "2017-04-16",
@@ -44,6 +46,7 @@ class TestPoland(TestCase):
 
     def test_2022(self):
         self.assertHolidayDates(
+            Poland(years=2022),
             "2022-01-01",
             "2022-01-06",
             "2022-04-17",
@@ -69,16 +72,16 @@ class TestPoland(TestCase):
         self.assertHoliday(f"{year}-01-06" for year in range(1925, 1960))
         self.assertHoliday(f"{year}-01-06" for year in range(2011, 2050))
         self.assertNoHoliday(f"{year}-01-06" for year in range(1961, 2011))
-        self.assertNoHolidayName(
-            "Święto Trzech Króli", Poland(years=range(1961, 2011))
+        self.assertNoHolidayNameInYears(
+            "Święto Trzech Króli", range(1961, 2011)
         )
 
     def test_oczyszczenie_nmp(self):
         self.assertHoliday(f"{year}-02-02" for year in range(1925, 1951))
         self.assertNoHoliday(f"{year}-02-02" for year in range(1951, 2050))
-        self.assertNoHolidayName(
+        self.assertNoHolidayNameInYears(
             "Oczyszczenie Najświętszej Marii Panny",
-            Poland(years=range(1951, 2050)),
+            range(1951, 2050),
         )
 
     def test_easter_related(self):
@@ -128,29 +131,27 @@ class TestPoland(TestCase):
     def test_swieto_panstwowe(self):
         self.assertHoliday(f"{year}-05-01" for year in range(1950, 2050))
         self.assertNoHoliday(f"{year}-05-01" for year in range(1925, 1950))
-        self.assertNoHolidayName(
-            "Święto Państwowe", Poland(years=range(1925, 1950))
-        )
+        self.assertNoHolidayNameInYears("Święto Państwowe", range(1925, 1950))
 
     def test_swieto_narodowe_trzeciego_maja(self):
         self.assertHoliday(f"{year}-05-03" for year in range(1925, 1951))
         self.assertHoliday(f"{year}-05-03" for year in range(1990, 2050))
         self.assertNoHoliday(f"{year}-05-03" for year in range(1951, 1990))
-        self.assertNoHolidayName(
-            "Święto Narodowe Trzeciego Maja", Poland(years=range(1951, 1990))
+        self.assertNoHolidayNameInYears(
+            "Święto Narodowe Trzeciego Maja", range(1951, 1990)
         )
 
     def test_narodowe_swieto_zwyciestwa_i_wolnosci(self):
         self.assertHoliday(f"{year}-05-09" for year in range(1946, 1951))
         self.assertNoHoliday(f"{year}-05-09" for year in range(1925, 1946))
         self.assertNoHoliday(f"{year}-05-09" for year in range(1951, 2050))
-        self.assertNoHolidayName(
+        self.assertNoHolidayNameInYears(
             "Narodowe Święto Zwycięstwa i Wolności",
-            Poland(years=range(1925, 1946)),
+            range(1925, 1946),
         )
-        self.assertNoHolidayName(
+        self.assertNoHolidayNameInYears(
             "Narodowe Święto Zwycięstwa i Wolności",
-            Poland(years=range(1951, 2050)),
+            range(1951, 2050),
         )
 
     def test_wniebowstapienie_panskie(self):
@@ -161,8 +162,8 @@ class TestPoland(TestCase):
             "1945-05-11",
             "1950-05-19",
         )
-        self.assertNoHolidayName(
-            "Wniebowstąpienie Pańskie", Poland(years=range(1951, 2050))
+        self.assertNoHolidayNameInYears(
+            "Wniebowstąpienie Pańskie", range(1951, 2050)
         )
 
     def test_drugi_dzien_zielonych_swiatek(self):
@@ -173,38 +174,38 @@ class TestPoland(TestCase):
             "1945-05-21",
             "1950-05-29",
         )
-        self.assertNoHolidayName(
-            "Drugi dzień Zielonych Świątek", Poland(years=range(1951, 2050))
+        self.assertNoHolidayNameInYears(
+            "Drugi dzień Zielonych Świątek", range(1951, 2050)
         )
 
     def test_swietych_apostolow_piotra_i_pawla(self):
         self.assertHoliday(f"{year}-06-29" for year in range(1925, 1951))
         self.assertNoHoliday(f"{year}-06-29" for year in range(1951, 2050))
-        self.assertNoHolidayName(
+        self.assertNoHolidayNameInYears(
             "Uroczystość Świętych Apostołów Piotra i Pawła",
-            Poland(years=range(1951, 2050)),
+            range(1951, 2050),
         )
 
     def test_narodowe_swieto_odrodzenia_polski(self):
         self.assertHoliday(f"{year}-07-22" for year in range(1945, 1990))
         self.assertNoHoliday(f"{year}-07-22" for year in range(1925, 1945))
         self.assertNoHoliday(f"{year}-07-22" for year in range(1990, 2050))
-        self.assertNoHolidayName(
+        self.assertNoHolidayNameInYears(
             "Narodowe Święto Odrodzenia Polski",
-            Poland(years=range(1925, 1945)),
+            range(1925, 1945),
         )
-        self.assertNoHolidayName(
+        self.assertNoHolidayNameInYears(
             "Narodowe Święto Odrodzenia Polski",
-            Poland(years=range(1990, 2050)),
+            range(1990, 2050),
         )
 
     def test_wniebowziecie_nmp(self):
         self.assertHoliday(f"{year}-08-15" for year in range(1925, 1961))
         self.assertHoliday(f"{year}-08-15" for year in range(1989, 2050))
         self.assertNoHoliday(f"{year}-08-15" for year in range(1961, 1989))
-        self.assertNoHolidayName(
+        self.assertNoHolidayNameInYears(
             "Wniebowzięcie Najświętszej Marii Panny",
-            Poland(years=range(1961, 1989)),
+            range(1961, 1989),
         )
 
     def test_wszystkich_swietych(self):
@@ -214,20 +215,20 @@ class TestPoland(TestCase):
         self.assertHoliday(f"{year}-11-11" for year in range(1937, 1945))
         self.assertHoliday(f"{year}-11-11" for year in range(1989, 2050))
         self.assertNoHoliday(f"{year}-11-11" for year in range(1925, 1937))
-        self.assertNoHoliday(f"{year}-11-11" for year in range(1925, 1937))
-        self.assertNoHolidayName(
-            "Narodowe Święto Niepodległości", Poland(years=range(1925, 1937))
+        self.assertNoHoliday(f"{year}-11-11" for year in range(1945, 1989))
+        self.assertNoHolidayNameInYears(
+            "Narodowe Święto Niepodległości", range(1925, 1937)
         )
-        self.assertNoHolidayName(
-            "Narodowe Święto Niepodległości", Poland(years=range(1925, 1937))
+        self.assertNoHolidayNameInYears(
+            "Narodowe Święto Niepodległości", range(1945, 1989)
         )
 
     def test_niepokalane_poczecie_nmp(self):
         self.assertHoliday(f"{year}-12-08" for year in range(1925, 1951))
         self.assertNoHoliday(f"{year}-12-08" for year in range(1951, 2050))
-        self.assertNoHolidayName(
+        self.assertNoHolidayNameInYears(
             "Niepokalane Poczęcie Najświętszej Marii Panny",
-            Poland(years=range(1951, 2050)),
+            range(1951, 2050),
         )
 
     def test_boze_narodzenie(self):