--- a/tests/countries/test_vietnam.py
+++ b/tests/countries/test_vietnam.py
@@ -18,8 +18,9 @@ from tests.common import TestCase
 
 class TestVietnam(TestCase):
     @classmethod
-    def setUpClass(self):
-        super().setUpClass(Vietnam)
+    def setUpClass(cls):
+        years = range(1979, 2050)
+        super().setUpClass(Vietnam, years=years)
 
     def test_country_aliases(self):
         self.assertCountryAliases(Vietnam, VN, VNM)
@@ -106,7 +107,7 @@ class TestVietnam(TestCase):
         )
 
     def test_observed(self):
-        self.assertHoliday(
+        observed_holidays = (
             # International New Year's Day.
             "2012-01-02",
             "2017-01-02",
@@ -134,8 +135,5 @@ class TestVietnam(TestCase):
             "2018-09-03",
             "2023-09-04",
         )
-
-    def test_not_observed(self):
-        self.assertNoNonObservedHoliday(
-            "2023-01-02",
-        )
+        self.assertHoliday(observed_holidays)
+        self.assertNoNonObservedHoliday(observed_holidays)